Add files via upload
Assignment 2
Memory Management

In this assignment, you are tasked with simulating a virtual memory management system.

Broadly speaking, your simulation will have the following responsibilities:
1) Simulate a paging system
2) Implement three different page replacement algorithms
3) Implement a variable page size
4) Implement demand/prepaging
5) Record page swaps during a run

Let's discuss each phase of this assignment in turn.

1: Simulate a paging system
This simulation will use the idea of a 'memory location' atomic unit.  The pages in our simulation will be expressed in terms of this idea.  Thus, if our page size is 2, we have two memory locations on each page.  

Main memory, in our program, will be 512 memory locations big. 

Supplied with this assignment are two files, programlist, which contains the list of programs that we will be loading into main memory, and programtrace, which contains a deterministic series of memory access that emulates a real systems me