Add files via upload
This assignment involves generating a simple client/server chatroom with sockets.  The server should support up to 10 clients at one time.  The client's behavior will be defined as the following: upon starting the client, the user will be asked to supply a hostname for the server, and a username. 

 Once the connection is formed, the chatroom behavior is defined as follows:
 A user types a message and presses return/enter.  This message is sent to the server, and echoed onto any other connected client terminals.   Beyond this basic behavior, a few special commands should be defined.  If a user types /exit, /quit, or /part, the client should exit. The server should then print a message that the client(identified by the username he gave) has left the room.  But if the user presses control-C at the client side, instead of exiting, the client should print a nice error message asking user to type /exit, /quit, or /part to exit the chatroom.

Finally,   the server should gracefully quit o